Something is not working when changing adc channel

+++ disco/strip.h
@@ -11,11 +11,12 @@
 
 #include <avr/io.h>
 #include "io.h"
+#include <stdlib.h>
 
 //Public functions
 void strip_setChannel(Color col,uint8_t val);
 void strip_setRGB(uint8_t r, uint8_t g, uint8_t b);
-void strip_setHSV(int hVal,int sVal, int vVal);
+void strip_setHSV(uint16_t hVal,uint16_t sVal, uint16_t vVal);
 void strip_setNewHSVColor();//returns new HSV color from colors 2x array
 //All fade-funcions uses HSV values and colors 2x array
 uint8_t strip_fade(uint8_t stepsize); //returns true if faded to fadecolor
